Colors file added, layout changed
