--- Navigation/FeedViewController.swift
@@ -1,42 +1,0 @@
-//
-//  feedViewController.swift
-//  Navigation
-//
-//  Created by V SEO on 28.05.2022.
-//
-
-import UIKit
-
-class FeedViewController: UIViewController {
-    
-//    private lazy var button: UIButton = {
-//        let button = UIButton()
-//        button.addTarget(self, action: #selector(hasBeenTapped), for: .touchUpInside)
-//        button.setTitle("Просмотр поста", for: .normal)
-//        return button
-//    }()
-//
-//    override func viewDidLoad() {
-//        super.viewDidLoad()
-//
-//        view.addSubview(button)
-//        button.translatesAutoresizingMaskIntoConstraints = false
-//        NSLayoutConstraint.activate([
-//            button.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-//            button.centerYAnchor.constraint(equalTo: view.centerYAnchor),
-//            button.heightAnchor.constraint(equalToConstant: 145),
-//            button.widthAnchor.constraint(equalToConstant: 150)
-//        ])
-//    }
-//
-//    @objc
-//    private func hasBeenTapped() {
-//        let post = Post(title: "Заголовок поста")
-//        let controller = PostViewController(post: post)
-//        navigationController?.pushViewController(controller, animated: true)
-        //let navVC = UINavigationController(rootViewController: PostViewController())
-        //present(navVC, animated: true)
-    // }
-   
-
-}
