--- Navigation/InfoViewController.swift
@@ -1,12 +1,0 @@
-//
-//  MainTabBar.swift
-//  Navigation
-//
-//  Created by V SEO on 28.05.2022.
-//
-
-import UIKit
-
-class InfoViewController: UIViewController {
-    
-}
