--- Navigation/ViewController.swift
@@ -1,21 +1,0 @@
-//
-//  ViewController.swift
-//  Navigation
-//
-//  Created by V SEO on 28.05.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .purple
-      //  setupControllers()
-    }
-
-  
-    
-}
-
